Update README.md
The code you provided is an example of an ESP32-based IoT project that measures temperature and humidity using a DHT22 sensor and sends the data to a server using HTTP POST requests. Here's a breakdown of the code:

Importing Required Libraries:

WiFi.h: Library for connecting to a Wi-Fi network.
HTTPClient.h: Library for making HTTP requests.
DHT.h: Library for interacting with DHT sensors.
Defining Constants:

DHTPIN: The GPIO pin number to which the DHT sensor is connected.
DHTTYPE: The type of DHT sensor being used (DHT22 in this case).
URL: The URL where the data will be sent.
Setting Wi-Fi Credentials:

ssid: The name (SSID) of the Wi-Fi network to connect to.
password: The password of the Wi-Fi network.
Initializing Variables:

temperature and humidity: Variables to store the measured temperature and humidity values.
setup() Function:

Initializes the serial communication for debugging.
Initializes the DHT sensor.
loop() Function:

Checks if the ESP32 is connected to the Wi-Fi network. If not, it reconnects.
Reads the temperature and humidity values from the DHT22 sensor.
Checks if the sensor reading was successful. If not, sets temperature and humidity to 0.
Prints the temperature and humidity values to the serial monitor.
Creates a string postData to store the data in the format expected by the server.
Creates an HTTPClient object and sends an HTTP POST request to the specified URL.
Checks the HTTP response code and prints the response payload or error message.
Closes the HTTP connection.
Prints various debug information to the serial monitor.
Delays for 5 seconds before repeating the process.
Overall, this code continuously reads temperature and humidity values from the DHT22 sensor, sends the data to the specified server URL using HTTP POST requests, and displays the response from the server on the serial monitor for debugging purposes.